requirements.txt
asgiref==3.2.10
Django==3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
docutils==0.17.1
mysqlclient==2.0.3
Pillow==8.3.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
